+++ resources/themes/dashcode/views/starter/profile/edit-my-profile.blade.php
@@ -111,7 +111,7 @@
             </aside>
         </div>
 
-        <div data-starter-region="section-content">
+        <div class="min-w-0" data-starter-region="section-content">
             <div class="tab-content">
                 <form
                     id="account-details"